backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
onnx==1.14.1
onnxruntime==1.16.0
xgboost==2.0.0
shap==0.43.0
requests==2.31.0
python-dotenv==1.0.0
tldextract==5.3.0